requirements.txt
pydantic>=2.0.0
pyyaml>=6.0
pandas>=2.0.0
pytest>=7.0.0
